Refactor Python logic into separate file
